ONNX import refactor checkpoint. Also performed some code cleaning in some functions
